+++ SKYAPI/SKYAPI.psm1
@@ -86,6 +86,11 @@ Function Get-SKYAPIAuthToken
     [CmdletBinding()]
     Param($grant_type,$client_id,$redirect_uri,$client_secret,$authCode,$token_uri)
 
+    # Disable Progress Bar in Function Scope When Calling Invoke-WebRequest or Invoke-RestMethod.
+    # This improves performance due to a bug in some versions of PowerShell. It was eventually fixed in Core (v6.0.0-alpha.13) but still is around in Desktop.
+    # More Information: https://github.com/PowerShell/PowerShell/pull/2640
+    $ProgressPreference = 'SilentlyContinue'
+
     #Build token request
     $AuthorizationPostRequest = 'grant_type=' + $grant_type + '&' +
     'redirect_uri=' + [System.Web.HttpUtility]::UrlEncode($redirect_uri) + '&' +
@@ -112,6 +117,11 @@ Function Get-SKYAPIAccessToken
     [CmdletBinding()]
     Param($grant_type,$client_id,$redirect_uri,$client_secret,$authCode,$token_uri)
 
+    # Disable Progress Bar in Function Scope When Calling Invoke-WebRequest or Invoke-RestMethod.
+    # This improves performance due to a bug in some versions of PowerShell. It was eventually fixed in Core (v6.0.0-alpha.13) but still is around in Desktop.
+    # More Information: https://github.com/PowerShell/PowerShell/pull/2640
+    $ProgressPreference = 'SilentlyContinue'
+    
     #Build token request
     $AuthorizationPostRequest = 'grant_type=' + $grant_type + '&' +
     'redirect_uri=' + [System.Web.HttpUtility]::UrlEncode($redirect_uri) + '&' +
@@ -173,55 +183,6 @@ function Resolve-SKYAPIMemberChain
     }
 }
 
-# Helper to make sure Browser Emulation/Compatibility Mode is Off When Using the WebBrowser Control.
-# This function will set the Internet Explorer emulation mode for the running executable. This allows the WebBrowser control to support newer html features and improves compatibility with modern websites.
-# Modified from https://www.sapien.com/blog/2020/11/05/a-simple-fix-for-problems-with-windows-forms-webbrowser/ (see also https://bchallis.wordpress.com/2020/10/17/problems-with-the-windows-forms-webbrowser-control-and-a-simple-way-to-fix-it/)
-function Set-SKYAPIWebBrowserEmulation
-{
-	param
-	(
-		[ValidateNotNullOrEmpty()]
-		[string]
-		$ExecutableName = [System.IO.Path]::GetFileName([System.Diagnostics.Process]::GetCurrentProcess().MainModule.FileName)
-	)
- 
-	#region Get IE Version
-	$valueNames = 'svcVersion', 'svcUpdateVersion', 'Version', 'W2kVersion'
- 
-	$version = 0;
-	for ($i = 0; $i -lt $valueNames.Length; $i++)
-	{
-		$objVal = [Microsoft.Win32.Registry]::GetValue('HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Internet Explorer', $valueNames[$i], '0')
-		$strVal = [System.Convert]::ToString($objVal)
-		if ($strVal)
-		{
-			$iPos = $strVal.IndexOf('.')
-			if ($iPos -gt 0)
-			{
-				$strVal = $strVal.Substring(0, $iPos)
-			}
- 
-			$res = 0;
-			if ([int]::TryParse($strVal, [ref]$res))
-			{
-				$version = [Math]::Max($version, $res)
-			}
-		}
-	}
- 
-	if ($version -lt 7)
-	{
-		$version = 7000
-	}
-	else
-	{
-		$version = $version * 1000
-	}
-	#endregion
- 
-	[Microsoft.Win32.Registry]::SetValue('HKEY_CURRENT_USER\SOFTWARE\Microsoft\Internet Explorer\Main\FeatureControl\FEATURE_BROWSER_EMULATION', $ExecutableName, $version)
-}
-
 Function Show-SKYAPIOAuthWindow
 {
     Param(
@@ -237,15 +198,15 @@ Function Show-SKYAPIOAuthWindow
         Mandatory=$false,
         ValueFromPipeline=$true,
         ValueFromPipelineByPropertyName=$true)]
-        [ValidateSet('','EdgeWebView2','MiniHTTPServer','LegacyIEControl')] # Allows null to be passed
+        [ValidateSet('','EdgeWebView2')] # Allows null to be passed
         [string]$AuthenticationMethod,
 
         [parameter(
-            Position=2,
-            Mandatory=$false,
-            ValueFromPipeline=$true,
-            ValueFromPipelineByPropertyName=$true)]
-            [switch]$ClearBrowserControlCache
+        Position=2,
+        Mandatory=$false,
+        ValueFromPipeline=$true,
+        ValueFromPipelineByPropertyName=$true)]
+        [switch]$ClearBrowserControlCache
     )
 
     # If Edge WebView 2 is the Authentication Method & the runtime not installed - https://developer.microsoft.com/en-us/microsoft-edge/webview2/
@@ -284,7 +245,7 @@ Function Show-SKYAPIOAuthWindow
             Write-Warning "Microsoft Edge WebView2 Runtime is not installed and is required for browser-based authentication. Please install the runtime and try again."
             $PromptNoWebView2Runtime_Title = "Options"
             $PromptNoWebView2Runtime_Message = "Enter your choice:"
-            $PromptNoWebView2Runtime_Choices = [System.Management.Automation.Host.ChoiceDescription[]]@("&Download & install the Edge WebView2 runtime", "&Try alternative method (beta)", "&Cancel & exit")
+            $PromptNoWebView2Runtime_Choices = [System.Management.Automation.Host.ChoiceDescription[]]@("&Download & install the Edge WebView2 runtime", "&Cancel & exit")
             $PromptNoWebView2Runtime_Default = 0
             $PromptNoWebView2Runtime_Selection = $host.UI.PromptForChoice($PromptNoWebView2Runtime_Title,$PromptNoWebView2Runtime_Message,$PromptNoWebView2Runtime_Choices,$PromptNoWebView2Runtime_Default)
 
@@ -323,9 +284,6 @@ Function Show-SKYAPIOAuthWindow
                         Write-Host "Retrying Authentication...`n"
                     }
                 1   {
-                        $AuthenticationMethod = "MiniHTTPServer"
-                    }
-                2   {
                         Write-Host "Exiting..."
                         Exit
                     }
@@ -335,85 +293,6 @@ Function Show-SKYAPIOAuthWindow
     
     switch ($AuthenticationMethod)
     {
-        MiniHTTPServer # TODO
-        {
-            Write-Host "`nUsing this option will attempt to authenticate using an alternate method by building a mini webserver in PowerShell. Continue?"
-            $PromptMiniWebserver_Title = "Options"
-            $PromptMiniWebserver_Message = "Enter your choice:"
-            $PromptMiniWebserver_Choices = [System.Management.Automation.Host.ChoiceDescription[]]@("&Load temporary HTTP server", "&Cancel & exit")
-            $PromptMiniWebserver_Default = 0
-            $PromptMiniWebserver_Selection = $host.UI.PromptForChoice($PromptMiniWebserver_Title,$PromptMiniWebserver_Message,$PromptMiniWebserver_Choices,$PromptMiniWebserver_Default)
-
-            switch($PromptMiniWebserver_Selection)
-            {
-                0   {
-                        Write-Warning "Sorry. The mini webserver authentication feature is not yet implemented."
-                        Write-Host "Exiting..."
-                        Exit
-                    }
-                1   {
-                        Write-Host "Exiting..."
-                        Exit
-                    }
-            }
-        }
-        LegacyIEControl
-        {
-            Set-SKYAPIWebBrowserEmulation
-
-            if ($ClearBrowserControlCache)
-            {
-                # Try to clear IE cache
-                # More info: https://superuser.com/questions/450014/clearmytracksbyprocess-all-options
-                # Using 4351 (0x10FF) to clear all + files and settings stored by add-ons. Convert Hex to Decimal.
-                # // This magic value is the combination of the following bitflags:
-                # // #define CLEAR_HISTORY         0x0001 // Clears history
-                # // #define CLEAR_COOKIES         0x0002 // Clears cookies
-                # // #define CLEAR_CACHE           0x0004 // Clears Temporary Internet Files folder
-                # // #define CLEAR_CACHE_ALL       0x0008 // Clears offline favorites and download history
-                # // #define CLEAR_FORM_DATA       0x0010 // Clears saved form data for form auto-fill-in
-                # // #define CLEAR_PASSWORDS       0x0020 // Clears passwords saved for websites
-                # // #define CLEAR_PHISHING_FILTER 0x0040 // Clears phishing filter data
-                # // #define CLEAR_RECOVERY_DATA   0x0080 // Clears webpage recovery data
-                # // #define CLEAR_PRIVACY_ADVISOR 0x0800 // Clears tracking data
-                # // #define CLEAR_SHOW_NO_GUI     0x0100 // Do not show a GUI when running the cache clearing
-                # //
-                # // Bitflags available but not used in this magic value are as follows:
-                # // #define CLEAR_USE_NO_THREAD      0x0200 // Do not use multithreading for deletion
-                # // #define CLEAR_PRIVATE_CACHE      0x0400 // Valid only when browser is in private browsing mode
-                # // #define CLEAR_DELETE_ALL         0x1000 // Deletes data stored by add-ons
-                # // #define CLEAR_PRESERVE_FAVORITES 0x2000 // Preserves cached data for "favorite" websites
-                Write-Warning "Note: You may have to close PowerShell and start a new session for clearing the IE cache to take effect."
-                Start-Process -FilePath 'RunDll32.exe' -ArgumentList 'InetCpl.cpl, ClearMyTracksByProcess 4351' -Wait
-                $ClearBrowserControlCache = $false
-            }
-
-            Add-Type -AssemblyName System.Windows.Forms
-        
-            $form = New-Object -TypeName System.Windows.Forms.Form -Property @{Width=600;Height=800}
-            $web = New-Object -TypeName System.Windows.Forms.WebBrowser -Property @{Width=584;Height=760;Url=($url)}
-            $DocComp = {
-                $Global:uri = $web.Url.AbsoluteUri
-                if ($Global:Uri -match "error=[^&]*|code=[^&]*") {$form.Close() }
-            }
-            $web.ScriptErrorsSuppressed = $true
-            $web.Add_DocumentCompleted($DocComp)
-
-            $form.Controls.Add($web)
-            $form.Add_Shown({$form.Activate()})
-            $form.ShowDialog() | Out-Null
-
-            # Parse Return URL
-            $queryOutput = [System.Web.HttpUtility]::ParseQueryString($web.Url.Query)
-            $output = @{}
-            foreach($key in $queryOutput.Keys){
-                $output["$key"] = $queryOutput[$key]
-            }
-
-            # Dispose Form & IE WebBrowser Control
-            $web.Dispose()
-            $form.Dispose()
-        }
         default # EdgeWebView2
         {            
             # Set EdgeWebView2 Control Version to Use
@@ -539,26 +418,26 @@ Function Get-SKYAPINewTokens
     [CmdletBinding()]
     Param(
         [parameter(
-            Position=0,
-            Mandatory=$false,
-            ValueFromPipeline=$true,
-            ValueFromPipelineByPropertyName=$true)]
-            [string]$sky_api_tokens_file_path,
+        Position=0,
+        Mandatory=$false,
+        ValueFromPipeline=$true,
+        ValueFromPipelineByPropertyName=$true)]
+        [string]$sky_api_tokens_file_path,
         
         [parameter(
-            Position=1,
-            Mandatory=$false,
-            ValueFromPipeline=$true,
-            ValueFromPipelineByPropertyName=$true)]
-            [ValidateSet('','EdgeWebView2','MiniHTTPServer','LegacyIEControl')] # Allows null to be passed
-            [string]$AuthenticationMethod,
+        Position=1,
+        Mandatory=$false,
+        ValueFromPipeline=$true,
+        ValueFromPipelineByPropertyName=$true)]
+        [ValidateSet('','EdgeWebView2')] # Allows null to be passed
+        [string]$AuthenticationMethod,
 
         [parameter(
-            Position=2,
-            Mandatory=$false,
-            ValueFromPipeline=$true,
-            ValueFromPipelineByPropertyName=$true)]
-            [switch]$ClearBrowserControlCache
+        Position=2,
+        Mandatory=$false,
+        ValueFromPipeline=$true,
+        ValueFromPipelineByPropertyName=$true)]
+        [switch]$ClearBrowserControlCache
     )
 
     # Set the Necessary Config Variables
@@ -600,9 +479,56 @@ Function Get-SKYAPINewTokens
         | Out-File -FilePath $sky_api_tokens_file_path -Force
 }
 
-# Handle Common Errors > https://developer.blackbaud.com/skyapi/docs/resources/in-depth-topics/handle-common-errors
-function SKYAPICatchInvokeErrors($InvokeErrorMessageRaw)
+function Get-ExponentialBackoffDelay
 {
+    [CmdletBinding()]
+    Param(
+        [parameter(
+        Position=0,
+        Mandatory=$true,
+        ValueFromPipeline=$true,
+        ValueFromPipelineByPropertyName=$true)]
+        [int]$InitialDelay,
+        
+        [parameter(
+        Position=1,
+        Mandatory=$true,
+        ValueFromPipeline=$true,
+        ValueFromPipelineByPropertyName=$true)]
+        [int]$InvokeCount 
+    )
+
+    # Return the delay time.
+    return ($InitialDelay * [Math]::Pow(2, $InvokeCount - 1)) # Initial delay times 2 to the power of $InvokeCount minus 1.
+}
+
+# Handle Common Errors > https://developer.blackbaud.com/skyapi/docs/in-depth-topics/handle-common-errors
+function SKYAPICatchInvokeErrors
+{
+    [CmdletBinding()]
+    Param(
+        [parameter(
+        Position=0,
+        Mandatory=$true,
+        ValueFromPipeline=$true,
+        ValueFromPipelineByPropertyName=$true)]
+        $InvokeErrorMessageRaw,
+        
+        [parameter(
+        Position=1,
+        Mandatory=$true,
+        ValueFromPipeline=$true,
+        ValueFromPipelineByPropertyName=$true)]
+        [int]$InvokeCount,
+
+        [parameter(
+        Position=2,
+        Mandatory=$true,
+        ValueFromPipeline=$true,
+        ValueFromPipelineByPropertyName=$true)]
+        [int]$MaxInvokeCount
+    )
+
     # Convert From JSON
     try
     {
@@ -637,64 +563,124 @@ function SKYAPICatchInvokeErrors($InvokeErrorMessageRaw)
     # Try and handle the error message.
     Switch ($StatusCodeorError)
     {
-        invalid_client # You usually see this error when providing an invalid .
+        invalid_client # You usually, but not always, see this error when providing an invalid client id.
         {
             # We will display the error, try again and handle the issue later.
             Write-Warning $InvokeErrorMessageRaw
+
+            # Check if we've hit the max invoke count and if so, throw the error.
+            if ($InvokeCount -ge $MaxInvokeCount)
+            {
+                throw $InvokeErrorMessageRaw
+            }
+
             'retry'
         }
         invalid_grant # You usually, but not always, see this error when providing an invalid, expired, or previously used authorization code.
         {
             # We will display the error, try again and handle the issue later.
             Write-Warning $InvokeErrorMessageRaw
+
+            # Check if we've hit the max invoke count and if so, throw the error.
+            if ($InvokeCount -ge $MaxInvokeCount)
+            {
+                throw $InvokeErrorMessageRaw
+            }
+
             'retry'
         }
-        400 # Bad request. Usually means that data in the initial request is invalid or improperly formatted.
+        400 # Bad Request. Usually means that data in the initial request is invalid or improperly formatted.
         {
             throw $InvokeErrorMessageRaw
         }
-        401 # Unauthorized Request. Could mean that the authenticated user does not have rights to access the requested data or does not have permission to edit a given record or record type. An unauthorized request also occurs if the authorization token expires or if the authorization header is not supplied.
+        401 # Unauthorized. Could mean that the authenticated user does not have rights to access the requested data or does not have permission to edit a given record or record type. An unauthorized request also occurs if the authorization token expires or if the authorization header is not supplied.
         {
+            # Check if we've hit the max invoke count and if so, throw the error.
+            if ($InvokeCount -ge $MaxInvokeCount)
+            {
+                throw $InvokeErrorMessageRaw
+            }
+            
             # This can happens if the token has expired so we will try to refresh and then run the invoke again.
             Connect-SKYAPI -ForceRefresh
             'retry'
         }
-        429 # Rate limit is exceeded. Try again in 1 seconds. Technically, the number of seconds is returned in the 'Retry-After' header, but I think it's best not to wait longer. 
+        403 # Forbidden. The request failed because the user in whose context the API is being called either does not have permission to perform the operation itself, or does not have permission to access the data being requested. You may also see this response when the API quota associated with your subscription has been met.
         {
-            # Sleep for 1 second and return the try command.
+            # In addition to 429 rate limits (per second limit), SKY API also employs a quota limit to manage API traffic over a broader period of time. If this this limit is reached, requests return the 403 (Forbidden) status code with retry-after headers that indicate how long to wait before retrying an API request. Similar to the 429 responses, it is recommended to wait and retry after the time period in the retry-after header.
+            # TODO: Check for '403 - Quota Exceeded' response from the API because this is a different type of 403 error and means the broad period (as opposed to per-second) quota is exceeded and not a real "Forbidden" error.
+            throw $InvokeErrorMessageRaw
+        }
+        404 # Not Found. The requested resource could not be found. You may be trying to access a record that does not exist, or you may have supplied an invalid URL.
+        {
+            throw $InvokeErrorMessageRaw
+        }
+        415 # Unsupported Media Type. The request failed because the correct Content-Type header was not provided on the request. For endpoints that accept JSON in the request body, you must use the Content-Type header application/json.
+        {
+            throw $InvokeErrorMessageRaw
+        }
+        429 # Too Many Requests. Rate limit is exceeded. Try again in 1 seconds. Technically, the number of seconds is returned in the 'Retry-After' header, but the standard throttle is 10 calls per second. See: https://developer.blackbaud.com/skyapi/docs/in-depth-topics/api-request-throttling
+        {
+            # Check if we've hit the max invoke count and if so, throw the error.
+            if ($InvokeCount -ge $MaxInvokeCount)
+            {
+                throw $InvokeErrorMessageRaw
+            }
+
+            # Sleep for 1 second and return the retry action command.
             Start-Sleep -Seconds 1
             'retry'
         }
-        500 # Internal Server Error.
+        500 # Internal Server Error. An unexpected error has occurred on the SKY API side. You should never receive this response, but if you do let Blackbaud Support know.
         {
-            # Sleep for 5 seconds and return the try command. I don't know if this is a good length, but it seems reasonable since we try 5 times before failing.
-            # The other option would be to use the exponential backoff method where You can periodically retry a failed request over an increasing amount of time to handle errors
-            # related to rate limits, network volume, or response time. For example, you might retry a failed request after one second, then after two seconds, and then after four seconds.
-            Start-Sleep -Seconds 5
+            # Check if we've hit the max invoke count and if so, throw the error.
+            if ($InvokeCount -ge $MaxInvokeCount)
+            {
+                throw $InvokeErrorMessageRaw
+            }
+
+            # Exponential backoff
+            $SleepTime = Get-ExponentialBackoffDelay -InitialDelay 5 -InvokeCount $InvokeCount
+            Start-Sleep -Seconds $SleepTime
             'retry'
         }
-        503 # The service is currently unavailable.
+        503 # Service Unavailable. The service is currently unavailable. One or more API services are not available. This is usually a temporary condition caused by an unexpected outage or due to planned downtime. Check the Issues page (https://status.blackbaud.com/?svcid=skydev) for more information.
         {
-            # Sleep for 5 seconds and return the try command. I don't know if this is a good length, but it seems reasonable since we try 5 times before failing.
-            # The other option would be to use the exponential backoff method where You can periodically retry a failed request over an increasing amount of time to handle errors
-            # related to rate limits, network volume, or response time. For example, you might retry a failed request after one second, then after two seconds, and then after four seconds.
-            Start-Sleep -Seconds 5
+            # Check if we've hit the max invoke count and if so, throw the error.
+            if ($InvokeCount -ge $MaxInvokeCount)
+            {
+                throw $InvokeErrorMessageRaw
+            }
+
+            # Exponential backoff
+            $SleepTime = Get-ExponentialBackoffDelay -InitialDelay 5 -InvokeCount $InvokeCount
+            Start-Sleep -Seconds $SleepTime
             'retry'
         }
         504 # Gateway Time-out.
         {
-            # Sleep for 5 seconds and return the try command. I don't know if this is a good length, but it seems reasonable since we try 5 times before failing.
-            # The other option would be to use the exponential backoff method where You can periodically retry a failed request over an increasing amount of time to handle errors
-            # related to rate limits, network volume, or response time. For example, you might retry a failed request after one second, then after two seconds, and then after four seconds.
-            Start-Sleep -Seconds 5
+            # Check if we've hit the max invoke count and if so, throw the error.
+            if ($InvokeCount -ge $MaxInvokeCount)
+            {
+                throw $InvokeErrorMessageRaw
+            }
+
+            # Exponential backoff
+            $SleepTime = Get-ExponentialBackoffDelay -InitialDelay 5 -InvokeCount $InvokeCount
+            Start-Sleep -Seconds $SleepTime
             'retry'
         }
-        'An exception occured. Please contact Support.' # Random exception. Often transient.
+        'An exception occurred. Please contact Support.' # Random exception. Often transient.
         {
-            # Sleep for 5 seconds and return the try command. I don't know if this is a good length, but it seems reasonable since we try 5 times before failing.
-            # The other option would be to use the exponential backoff method where You can periodically retry a failed request over an increasing amount of time to handle errors
-            # related to rate limits, network volume, or response time. For example, you might retry a failed request after one second, then after two seconds, and then after four seconds.
-            Start-Sleep -Seconds 5
+            # Check if we've hit the max invoke count and if so, throw the error.
+            if ($InvokeCount -ge $MaxInvokeCount)
+            {
+                throw $InvokeErrorMessageRaw
+            }
+
+            # Exponential backoff
+            $SleepTime = Get-ExponentialBackoffDelay -InitialDelay 5 -InvokeCount $InvokeCount
+            Start-Sleep -Seconds $SleepTime
             'retry'
         }
         default
@@ -737,9 +723,14 @@ Function Get-SKYAPIUnpagedEntity
         $Request.Query = $params.ToString()
     }
     
+    # Disable Progress Bar in Function Scope When Calling Invoke-WebRequest or Invoke-RestMethod.
+    # This improves performance due to a bug in some versions of PowerShell. It was eventually fixed in Core (v6.0.0-alpha.13) but still is around in Desktop.
+    # More Information: https://github.com/PowerShell/PowerShell/pull/2640
+    $ProgressPreference = 'SilentlyContinue'
+
     # Run Invoke Command and Catch Responses
     [int]$InvokeCount = 0
-    [int]$MaxInvokeCount = 5
+    [int]$MaxInvokeCount = 7
     do
     {      
         $InvokeCount += 1
@@ -784,7 +775,7 @@ Function Get-SKYAPIUnpagedEntity
         {
             # Process Invoke Error
             $LastCaughtError = ($_)
-            $NextAction = SKYAPICatchInvokeErrors($_)
+            $NextAction = SKYAPICatchInvokeErrors -InvokeErrorMessageRaw $_ -InvokeCount $InvokeCount -MaxInvokeCount $MaxInvokeCount
 
             # Just in case the token was refreshed by the error catcher, update these
             $AuthTokensFromFile = Get-SKYAPIAuthTokensFromFile
@@ -836,9 +827,14 @@ Function Get-SKYAPIPagedEntity
     # Create records array
     $allRecords = @()
 
+    # Disable Progress Bar in Function Scope When Calling Invoke-WebRequest or Invoke-RestMethod.
+    # This improves performance due to a bug in some versions of PowerShell. It was eventually fixed in Core (v6.0.0-alpha.13) but still is around in Desktop.
+    # More Information: https://github.com/PowerShell/PowerShell/pull/2640
+    $ProgressPreference = 'SilentlyContinue'
+
     # Run Invoke Command and Catch Responses
     [int]$InvokeCount = 0
-    [int]$MaxInvokeCount = 5
+    [int]$MaxInvokeCount = 7
     do
     {      
         $InvokeCount += 1
@@ -917,7 +913,7 @@ Function Get-SKYAPIPagedEntity
         {
             # Process Invoke Error
             $LastCaughtError = ($_)
-            $NextAction = SKYAPICatchInvokeErrors($_)
+            $NextAction = SKYAPICatchInvokeErrors -InvokeErrorMessageRaw $_ -InvokeCount $InvokeCount -MaxInvokeCount $MaxInvokeCount
 
             # Just in case the token was refreshed by the error catcher, update these
             $AuthTokensFromFile = Get-SKYAPIAuthTokensFromFile
@@ -958,10 +954,15 @@ Function Remove-SKYAPIEntity
     if ($null -ne $params -and $params -ne '') {
         $Request.Query = $params.ToString()
     }
-    
+
+    # Disable Progress Bar in Function Scope When Calling Invoke-WebRequest or Invoke-RestMethod.
+    # This improves performance due to a bug in some versions of PowerShell. It was eventually fixed in Core (v6.0.0-alpha.13) but still is around in Desktop.
+    # More Information: https://github.com/PowerShell/PowerShell/pull/2640
+    $ProgressPreference = 'SilentlyContinue'
+
     # Run Invoke Command and Catch Responses
     [int]$InvokeCount = 0
-    [int]$MaxInvokeCount = 5
+    [int]$MaxInvokeCount = 7
     do
     {      
         $InvokeCount += 1
@@ -991,7 +992,7 @@ Function Remove-SKYAPIEntity
         {
             # Process Invoke Error
             $LastCaughtError = ($_)
-            $NextAction = SKYAPICatchInvokeErrors($_)
+            $NextAction = SKYAPICatchInvokeErrors -InvokeErrorMessageRaw $_ -InvokeCount $InvokeCount -MaxInvokeCount $MaxInvokeCount
 
             # Just in case the token was refreshed by the error catcher, update these
             $AuthTokensFromFile = Get-SKYAPIAuthTokensFromFile
@@ -1032,9 +1033,14 @@ function Submit-SKYAPIEntity
     # Build Body
     $PostRequest = $params | ConvertTo-Json
 
+    # Disable Progress Bar in Function Scope When Calling Invoke-WebRequest or Invoke-RestMethod.
+    # This improves performance due to a bug in some versions of PowerShell. It was eventually fixed in Core (v6.0.0-alpha.13) but still is around in Desktop.
+    # More Information: https://github.com/PowerShell/PowerShell/pull/2640
+    $ProgressPreference = 'SilentlyContinue'
+
     # Run Invoke Command and Catch Responses
     [int]$InvokeCount = 0
-    [int]$MaxInvokeCount = 5
+    [int]$MaxInvokeCount = 7
     do
     {      
         $InvokeCount += 1
@@ -1065,7 +1071,7 @@ function Submit-SKYAPIEntity
         {
             # Process Invoke Error
             $LastCaughtError = ($_)
-            $NextAction = SKYAPICatchInvokeErrors($_)
+            $NextAction = SKYAPICatchInvokeErrors -InvokeErrorMessageRaw $_ -InvokeCount $InvokeCount -MaxInvokeCount $MaxInvokeCount
 
             # Just in case the token was refreshed by the error catcher, update these
             $AuthTokensFromFile = Get-SKYAPIAuthTokensFromFile
@@ -1106,9 +1112,14 @@ function Update-SKYAPIEntity
     # Build Body
     $PatchRequest = $params | ConvertTo-Json
 
+    # Disable Progress Bar in Function Scope When Calling Invoke-WebRequest or Invoke-RestMethod.
+    # This improves performance due to a bug in some versions of PowerShell. It was eventually fixed in Core (v6.0.0-alpha.13) but still is around in Desktop.
+    # More Information: https://github.com/PowerShell/PowerShell/pull/2640
+    $ProgressPreference = 'SilentlyContinue'
+
     # Run Invoke Command and Catch Responses
     [int]$InvokeCount = 0
-    [int]$MaxInvokeCount = 5
+    [int]$MaxInvokeCount = 7
     do
     {      
         $InvokeCount += 1
@@ -1139,7 +1150,7 @@ function Update-SKYAPIEntity
         {
             # Process Invoke Error
             $LastCaughtError = ($_)
-            $NextAction = SKYAPICatchInvokeErrors($_)
+            $NextAction = SKYAPICatchInvokeErrors -InvokeErrorMessageRaw $_ -InvokeCount $InvokeCount -MaxInvokeCount $MaxInvokeCount
 
             # Just in case the token was refreshed by the error catcher, update these
             $AuthTokensFromFile = Get-SKYAPIAuthTokensFromFile